Sixteen more queries the server would not accept
All arithmetic or assignment on a parameter Postgres cannot type, and
all fatal on every request.

`-$3` to record a refund, `$4 * $3` to reprice an invoice line,
`ABS($4 - $3)` to decide whether a cash drawer needs sign-off,
`CURRENT_DATE + $5` to place a roster shift and an appointment,
`($2 * 3) || ' days'` to bucket slow-moving stock, `CURRENT_DATE - $4`
to select a marketing cohort -- an untyped parameter has no operator, so
each raised rather than returning a number.

The rest are `COALESCE($n, <literal>)` against a typed column, which
resolves as text: visiting-hours capacity into an integer, a discharge
barrier's date, a diet order's enum, and two jsonb columns on an
adverse-drug-reaction report. And a `gender` enum compared with text
when screening clinical-trial candidates.

The ADR insert had two jsonb columns wrong and only reported one,
because Postgres stops at the first. Fixing `seriousness_criteria`
revealed `concomitant_drugs` on the next run. That is the third time
this has happened, so it is worth saying plainly: the baseline count
falls more slowly than the number of fixes, and a query is not clean
until the checker says so twice.

Two findings were the checker's fault, not the code's. The hospital-KPI
select is a `const &str` and the no-show report is a `String::from`,
both of which have their GROUP BY and remaining filters pushed on at run
time. Handing the opening half to the server reports a query nobody
sends. Both extractors now skip a literal that is bound to a `const
&str` or opened with `String::from(`.

215 to 199.

+++ medbrains/crates/medbrains-marketing/src/cohorts.rs
@@ -207,7 +207,7 @@ pub async fn create_clinical_cohort(
            AND ($3::uuid IS NULL OR e.department_id = $3) \
          GROUP BY c.id \
          HAVING max(e.encounter_date) IS NULL \
-             OR max(e.encounter_date) < (CURRENT_DATE - $4) \
+             OR max(e.encounter_date) < (CURRENT_DATE - $4::int) \
          LIMIT $5 \
          ON CONFLICT (tenant_id, cohort_id, contact_id) DO NOTHING",
     )
